Refactor into components and external data
