+++ src/App.js
@@ -1,5 +1,9 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import { Search, Star, Camera, CheckCircle, User, TrendingUp, Calendar, Filter, Plus, X, ChevronDown, LogOut, Sparkles, MapPin, Globe, Heart, MessageCircle, Award, Clock, ArrowLeft } from 'lucide-react';
+import { initialSampleReviews } from './data/sampleReviews';
+import Header from './components/Header';
+import ItemList from './components/ItemList';
+import ItemDetailPage from './components/ItemDetailPage';
 
 // --- Helper: Google Icon SVG ---
 const GoogleIcon = () => (
@@ -33,42 +37,6 @@ if (typeof document !== 'undefined' && !document.getElementById('scrollbar-style
 
 // Mock Firebase for preview environment
 // In a real application, you would initialize Firebase here and use its actual SDK methods.
-const initialSampleReviews = [
-    // Original 9 Items (imageUrl removed, will be generated, fallback added)
-    { id: "sample-1", itemName: "Local Car Repair", mainCategory: "Real World", category: "Local Services", subCategory: "Mechanics", location: { city: "Bangkok", district: "Lat Phrao" }, rating: 4.5, reviewCount: 23, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Car+Repair", reviews: [{ id: 1, author: "Sarah K.", rating: 5, date: "2024-03-15", comment: "Quick and honest car repair.", verified: true, helpful: 15 }] },
-    { id: "sample-2", itemName: "@bangkok_fashion_daily", mainCategory: "Online", category: "Instagram", subCategory: "Influencer", location: { city: "Online" }, rating: 4.2, reviewCount: 60, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Fashion+Blog", reviews: [{ id: 3, author: "Style Seeker", rating: 5, date: "2024-03-20", comment: "Great styling tips!", verified: true, helpful: 32 }] },
-    { id: "sample-3", itemName: "Pro Moving Services", mainCategory: "Real World", category: "Local Services", subCategory: "Movers", location: { city: "Bangkok", district: "Yan Nawa" }, rating: 4.9, reviewCount: 15, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=Moving+Services", reviews: [{ id: 5, author: "Customer C", rating: 5, date: "2024-06-10", comment: "Quick, efficient, and careful.", verified: true, helpful: 7 }] },
-    { id: "sample-4", itemName: "Authentic Isaan Restaurant", mainCategory: "Real World", category: "Restaurants & Cafes", subCategory: "Thai Food", location: { city: "Bangkok", district: "Bang Rak" }, rating: 4.7, reviewCount: 50, fallbackImageUrl: "https://placehold.co/400x250/f39c12/ffffff?text=Isaan+Restaurant", reviews: [{ id: 7, author: "FoodieBKK", rating: 5, date: "2024-05-10", comment: "Amazing Som Tum and Larb Gai!", verified: true, helpful: 25 }] },
-    { id: "sample-5", itemName: "Math Tutor Expert", mainCategory: "Real World", category: "Education Centers", subCategory: "Tutoring Center", location: { city: "Bangkok", district: "Pathum Wan" }, rating: 4.5, reviewCount: 10, fallbackImageUrl: "https://placehold.co/400x250/9b59b6/ffffff?text=Math+Tutor", reviews: [{ id: 8, author: "Student Parent", rating: 5, date: "2024-05-20", comment: "Our son's grades improved significantly.", verified: true, helpful: 5 }] },
-    { id: "sample-6", itemName: "Vintage Camera Shop", mainCategory: "Real World", category: "Goods & Products", subCategory: "Collectibles", location: { city: "Bangkok", district: "Bang Rak" }, rating: 4.0, reviewCount: 12, fallbackImageUrl: "https://placehold.co/400x250/1abc9c/ffffff?text=Camera+Shop", reviews: [{ id: 9, author: "CollectorGuy", rating: 4, date: "2024-06-01", comment: "Found a rare vintage camera.", verified: false, helpful: 3 }] },
-    { id: "sample-7", itemName: "Online Marketing Agency", mainCategory: "Online", category: "Website", subCategory: "Service Provider", location: { city: "Online" }, rating: 4.7, reviewCount: 8, fallbackImageUrl: "https://placehold.co/400x250/34495e/ffffff?text=Marketing+Agency", reviews: [{ id: 10, author: "Client M", rating: 5, date: "2024-06-15", comment: "Amazing post-processing work.", verified: true, helpful: 2 }] },
-    { id: "sample-8", itemName: "Hip Cafe", mainCategory: "Real World", category: "Restaurants & Cafes", subCategory: "Cafe", location: { city: "Bangkok", district: "Watthana" }, rating: 4.3, reviewCount: 30, fallbackImageUrl: "https://placehold.co/400x250/95a5a6/ffffff?text=Hip+Cafe", reviews: [{ id: 11, author: "Coffee Lover", rating: 5, date: "2024-06-18", comment: "Best pour-over coffee in Thong Lo.", verified: true, helpful: 10 }] },
-    { id: "sample-9", itemName: "Co-working Space", mainCategory: "Real World", category: "Local Services", subCategory: "Other", location: { city: "Bangkok", district: "Watthana" }, rating: 4.6, reviewCount: 25, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Co-working", reviews: [{ id: 12, author: "Digital Nomad", rating: 5, date: "2024-06-20", comment: "Fast internet, comfy chairs.", verified: true, helpful: 8 }] },
-
-    // --- NEWLY ADDED ITEMS FOR FULL SUB-CATEGORY COVERAGE ---
-    { id: "new-fb-shop", itemName: "BKK Gadgets (FB)", mainCategory: "Online", category: "Facebook", subCategory: "Shop", location: { city: "Online" }, rating: 4.8, reviewCount: 150, fallbackImageUrl: "https://placehold.co/400x250/2980b9/ffffff?text=Gadgets+Shop", reviews: [{ id: 13, author: "TechieTom", rating: 5, date: "2024-06-12", comment: "Fast shipping on Facebook.", verified: true, helpful: 45 }] },
-    { id: "new-fb-page", itemName: "Thai Food Lovers Page", mainCategory: "Online", category: "Facebook", subCategory: "Page", location: { city: "Online" }, rating: 4.9, reviewCount: 1200, fallbackImageUrl: "https://placehold.co/400x250/8e44ad/ffffff?text=Food+Page", reviews: [{ id: 14, author: "Somchai C.", rating: 5, date: "2024-05-30", comment: "Best recipes.", verified: true, helpful: 250 }] },
-    { id: "new-fb-group", itemName: "Bangkok Expat Buy/Sell", mainCategory: "Online", category: "Facebook", subCategory: "Group", location: { city: "Online" }, rating: 4.5, reviewCount: 45, fallbackImageUrl: "https://placehold.co/400x250/2c3e50/ffffff?text=Buy/Sell+Group", reviews: [{ id: 15, author: "NewbieNancy", rating: 4, date: "2024-06-10", comment: "Found great deals.", verified: false, helpful: 12 }] },
-    { id: "new-shopee-shop", itemName: "SuperDeals Shopee", mainCategory: "Online", category: "Shopee", subCategory: "Shop", location: { city: "Online" }, rating: 4.7, reviewCount: 5000, fallbackImageUrl: "https://placehold.co/400x250/f1c40f/ffffff?text=Shopee+Shop", reviews: [{ id: 16, author: "ShopperSue", rating: 5, date: "2024-06-14", comment: "My go-to shop on Shopee.", verified: true, helpful: 123 }] },
-    { id: "new-shopee-seller", itemName: "Seller: JaneD_bkk", mainCategory: "Online", category: "Shopee", subCategory: "Seller", location: { city: "Online" }, rating: 4.9, reviewCount: 210, fallbackImageUrl: "https://placehold.co/400x250/e67e22/ffffff?text=Shopee+Seller", reviews: [{ id: 17, author: "HappyBuyer", rating: 5, date: "2024-06-11", comment: "Seller was very responsive.", verified: true, helpful: 30 }] },
-    { id: "new-ig-shop", itemName: "@ChicBoutique.BKK", mainCategory: "Online", category: "Instagram", subCategory: "Shop", location: { city: "Online" }, rating: 4.6, reviewCount: 88, fallbackImageUrl: "https://placehold.co/400x250/16a085/ffffff?text=Chic+Boutique", reviews: [{ id: 18, author: "FashionistaFae", rating: 5, date: "2024-06-05", comment: "The dress is gorgeous!", verified: true, helpful: 22 }] },
-    { id: "new-tiktok-influencer", itemName: "DanceMachine_TikTok", mainCategory: "Online", category: "TikTok", subCategory: "Influencer", location: { city: "Online" }, rating: 4.9, reviewCount: 1200, fallbackImageUrl: "https://placehold.co/400x250/27ae60/ffffff?text=TikTok+Influencer", reviews: [{ id: 19, author: "Fan1", rating: 5, date: "2025-06-17", comment: "Best dance challenges!", verified: false, helpful: 300 }] },
-    { id: "new-tiktok-shop", itemName: "TikTok Finds Store", mainCategory: "Online", category: "TikTok", subCategory: "Shop", location: { city: "Online" }, rating: 4.5, reviewCount: 120, fallbackImageUrl: "https://placehold.co/400x250/d35400/ffffff?text=TikTok+Shop", reviews: [{ id: 20, author: "ViralBuyer", rating: 4, date: "2024-06-30", comment: "Found the viral gadget here.", verified: true, helpful: 50 }] },
-    { id: "new-yt-channel", itemName: "Tech Tomorrow", mainCategory: "Online", category: "YouTube", subCategory: "Channel", location: { city: "Online" }, rating: 4.8, reviewCount: 450, fallbackImageUrl: "https://placehold.co/400x250/c0392b/ffffff?text=YouTube+Channel", reviews: [{ id: 21, author: "GadgetGeek", rating: 5, date: "2024-06-28", comment: "Unbiased and detailed reviews.", verified: true, helpful: 150 }] },
-    { id: "new-yt-creator", itemName: "Creator: Chef Praiya", mainCategory: "Online", category: "YouTube", subCategory: "Content Creator", location: { city: "Online" }, rating: 4.9, reviewCount: 800, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Chef+Creator", reviews: [{ id: 22, author: "HomeCook", rating: 5, date: "2024-06-25", comment: "Her cooking is so inspiring.", verified: true, helpful: 200 }] },
-    { id: "new-ls-tutor", itemName: "English Express Tutors", mainCategory: "Real World", category: "Local Services", subCategory: "Tutors", location: { city: "Bangkok", district: "Pathum Wan" }, rating: 4.7, reviewCount: 35, fallbackImageUrl: "https://placehold.co/400x250/1abc9c/ffffff?text=English+Tutor", reviews: [{ id: 23, author: "Parent_P", rating: 5, date: "2024-06-20", comment: "My daughter's confidence in English has soared.", verified: true, helpful: 10 }] },
-    { id: "new-ls-handyman", itemName: "Mr. Fix-It", mainCategory: "Real World", category: "Local Services", subCategory: "Handymen", location: { city: "Bangkok", district: "Watthana" }, rating: 4.8, reviewCount: 42, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=Handyman", reviews: [{ id: 24, author: "HarryH", rating: 5, date: "2024-06-01", comment: "Fixed my leaky faucet in no time.", verified: true, helpful: 18 }] },
-    { id: "new-ls-cleaner", itemName: "SparkleClean BKK", mainCategory: "Real World", category: "Local Services", subCategory: "Cleaners", location: { city: "Bangkok", district: "Sathorn" }, rating: 4.9, reviewCount: 75, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Cleaner", reviews: [{ id: 25, author: "BusyBob", rating: 5, date: "2024-06-15", comment: "My condo has never been this clean.", verified: true, helpful: 25 }] },
-    { id: "new-ls-plumber", itemName: "Reliable Plumbers", mainCategory: "Real World", category: "Local Services", subCategory: "Plumbers", location: { city: "Bangkok", district: "Chatuchak" }, rating: 4.7, reviewCount: 33, fallbackImageUrl: "https://placehold.co/400x250/9b59b6/ffffff?text=Plumber", reviews: [{ id: 26, author: "StressedSteve", rating: 5, date: "2024-05-25", comment: "Came at 2 AM for a burst pipe. Lifesavers!", verified: true, helpful: 11 }] },
-    { id: "new-ls-electrician", itemName: "Watt's Up Electricians", mainCategory: "Real World", category: "Local Services", subCategory: "Electricians", location: { city: "Bangkok", district: "Bang Kapi" }, rating: 4.8, reviewCount: 29, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Electrician", reviews: [{ id: 27, author: "OfficeOlly", rating: 5, date: "2024-06-02", comment: "Rewired our office over the weekend.", verified: true, helpful: 9 }] },
-    { id: "new-fb-car", itemName: "John's Used Cars (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Cars", location: { city: "Bangkok", district: "Bang Khen" }, rating: 4.4, reviewCount: 19, fallbackImageUrl: "https://placehold.co/400x250/f39c12/ffffff?text=Used+Car", reviews: [{ id: 28, author: "FrankF", rating: 4, date: "2024-06-08", comment: "Car was as described. Fair deal.", verified: false, helpful: 5 }] },
-    { id: "new-fb-furniture", itemName: "Secondhand Sofa (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Furniture", location: { city: "Bangkok", district: "Prawet" }, rating: 5.0, reviewCount: 1, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Used+Sofa", reviews: [{ id: 29, author: "HappyHome", rating: 5, date: "2025-06-17", comment: "Great condition and a great price!", verified: false, helpful: 1 }] },
-    { id: "new-fb-electronics", itemName: "Used iPhone (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Electronics", location: { city: "Bangkok", district: "Don Mueang" }, rating: 4.0, reviewCount: 5, fallbackImageUrl: "https://placehold.co/400x250/2c3e50/ffffff?text=Used+iPhone", reviews: [{ id: 30, author: "BuyerBen", rating: 4, date: "2024-06-29", comment: "Battery health was lower than stated, but works.", verified: false, helpful: 2 }] },
-    { id: "new-ll-condo", itemName: "Landlord: Mrs. Somchai", mainCategory: "Real World", category: "Landlords", subCategory: "Condo", location: { city: "Bangkok", district: "Huai Khwang" }, rating: 4.9, reviewCount: 8, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Condo+Building", reviews: [{ id: 31, author: "TenantTom", rating: 5, date: "2024-06-01", comment: "Best landlord ever.", verified: true, helpful: 4 }] },
-    { id: "new-ll-apt", itemName: "Sukhumvit Suites", mainCategory: "Real World", category: "Landlords", subCategory: "Apartment", location: { city: "Bangkok", district: "Khlong Toei" }, rating: 4.1, reviewCount: 22, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Apartment+Building", reviews: [{ id: 32, author: "ExpatEve", rating: 4, date: "2024-05-15", comment: "Good location, but the walls are a bit thin.", verified: true, helpful: 6 }] },
-    { id: "new-ll-house", itemName: "Rental House Ladprao", mainCategory: "Real World", category: "Landlords", subCategory: "House", location: { city: "Bangkok", district: "Lat Phrao" }, rating: 4.6, reviewCount: 3, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=House+Rental", reviews: [{ id: 33, author: "FamilyMan", rating: 5, date: "2024-04-10", comment: "Perfect for our family, great garden.", verified: true, helpful: 1 }] },
-];
-
 // MOCK USER DATA
 const mockUser = {
     uid: 'mock-google-user-123',
@@ -179,481 +147,8 @@ const generateImage = async (prompt) => {
 };
 
 
-// StarRating Component with enhanced design
-const StarRating = ({ rating, size = 'sm', interactive = false, onRate }) => {
-    const starSize = size === 'sm' ? 16 : size === 'md' ? 20 : 24;
-    const [hoveredStar, setHoveredStar] = useState(null);
-    
-    return (
-        <div className="flex items-center gap-0.5">
-            {[1, 2, 3, 4, 5].map((star) => (
-                <button
-                    key={star}
-                    onClick={() => interactive && onRate && onRate(star)}
-                    onMouseEnter={() => interactive && setHoveredStar(star)}
-                    onMouseLeave={() => interactive && setHoveredStar(null)}
-                    className={`transition-all duration-200 ${interactive ? 'cursor-pointer hover:scale-110' : 'cursor-default'}`}
-                    disabled={!interactive}
-                >
-                    <Star
-                        size={starSize}
-                        className={`transition-colors duration-200 ${
-                            (hoveredStar ? star <= hoveredStar : star <= rating)
-                                ? 'fill-amber-400 text-amber-400 drop-shadow-lg' 
-                                : 'text-gray-600 hover:text-gray-500'
-                        }`}
-                    />
-                </button>
-            ))}
-        </div>
-    );
-};
 
-// Enhanced ReviewItem Component
-const ReviewItem = ({ item, onClick, language, categories, citiesData, bangkokStreetsData, generatedImages, setGeneratedImages }) => {
-    const [isHovered, setIsHovered] = useState(false);
-    
-    // Get the specific image status for *this* item from the overall state
-    const currentItemImageStatus = generatedImages[item.id];
 
-    // Effect to generate image if not already generated or if it failed previously
-    useEffect(() => {
-        // Only generate if no image status or if there was an error in a previous attempt (up to 3 retries)
-        if (!currentItemImageStatus || (!currentItemImageStatus.url && !currentItemImageStatus.loading && (currentItemImageStatus.error || (currentItemImageStatus.retries !== undefined && currentItemImageStatus.retries < 3)))) {
-            const currentRetries = currentItemImageStatus?.retries || 0;
-            
-            // Only initiate generation if not already loading
-            if (currentItemImageStatus && currentItemImageStatus.loading) {
-                return;
-            }
-
-            setGeneratedImages(prev => ({ 
-                ...prev, 
-                [item.id]: { loading: true, url: null, error: false, retries: currentRetries + 1 } 
-            }));
-
-            // Modified prompt to specify English text only
-            const prompt = `A shop front image of "${item.itemName}" which is a ${item.subCategory || item.category} in ${item.location.city !== 'Online' ? item.location.city : 'digital space'}. Ensure any text visible in the image is in English. Style: realistic, high detail.`;
-            generateImage(prompt).then(url => {
-                setGeneratedImages(prev => ({ ...prev, [item.id]: { ...prev[item.id], loading: false, url: url, error: false } }));
-            }).catch(error => {
-                console.error("Failed to generate image for item:", item.itemName, error);
-                setGeneratedImages(prev => ({ ...prev, [item.id]: { ...prev[item.id], loading: false, url: null, error: true } })); 
-            });
-        }
-    }, [
-        item.id, 
-        item.itemName, 
-        item.category, 
-        item.subCategory, 
-        item.location.city, 
-        currentItemImageStatus?.url, 
-        currentItemImageStatus?.loading, 
-        currentItemImageStatus?.error, 
-        currentItemImageStatus?.retries, 
-        setGeneratedImages 
-    ]);
-
-    const getCategoryLabel = (mainCat, cat, subCat) => {
-        const mainLabel = categories.find(c => c.value === mainCat)?.[`label_${language}`] || mainCat;
-        const platformOrCat = mainCat === 'Online' 
-            ? categories.find(c => c.value === mainCat)?.platforms?.find(p => p.value === cat)?.[`label_${language}`] || cat
-            : categories.find(c => c.value === mainCat)?.subcategories?.find(s => s.value === cat)?.[`label_${language}`] || cat;
-        const subCatLabel = (mainCat === 'Online' ? (categories.find(c => c.value === mainCat)?.platforms?.find(p => p.value === cat)?.subcategories?.find(sc => sc.value === subCat)?.[`label_${language}`] || subCat) : (categories.find(c => c.value === mainCat)?.subcategories?.find(s => s.value === cat)?.subcategories.find(sc => sc.value === subCat)?.[`label_${language}`] || subCat));
-        
-        let parts = [mainLabel];
-        if (platformOrCat && platformOrCat !== cat) parts.push(platformOrCat);
-        if (subCatLabel && subCatLabel !== subCat) parts.push(subCatLabel);
-        
-        if (parts.length === 1 && cat && cat !== 'all') {
-            const currentCategoryData = (mainCat === 'Online' ? categories.find(c => c.value === mainCat)?.platforms : categories.find(c => c.value === mainCat)?.subcategories)?.find(c => c.value === cat);
-            if (currentCategoryData) parts.push(currentCategoryData[`label_${language}`] || cat);
-        }
-        if (parts.length === 2 && subCat && subCat !== 'all') {
-            let currentSubCategoryData;
-            if (mainCat === 'Online') {
-                const platform = categories.find(c => c.value === mainCat)?.platforms.find(p => p.value === cat);
-                currentSubCategoryData = platform?.subcategories.find(s => s.value === subCat);
-            } else {
-                const category = categories.find(c => c.value === mainCat)?.subcategories.find(s => s.value === cat);
-                currentSubCategoryData = category?.subcategories.find(sc => sc.value === subCat);
-            }
-            if (currentSubCategoryData) parts.push(currentSubCategoryData[`label_${language}`] || subCat);
-        }
-
-        return parts.filter(Boolean).join(' / ');
-    };
-
-    const getLocationLabel = (loc) => {
-        if (!loc || !loc.city) return '';
-        const cityLabel = citiesData.find(c => c.value === loc.city)?.[`label_${language}`] || loc.city;
-        const districtLabel = loc.district ? (citiesData.find(c => c.value === loc.city)?.districts.find(d => d.value === loc.district)?.[`label_${language}`] || loc.district) : '';
-        
-        let parts = [cityLabel];
-        if (districtLabel) parts.push(districtLabel);
-
-        return parts.filter(Boolean).join(', ');
-    };
-
-    const imageUrlToDisplay = currentItemImageStatus?.url || item.fallbackImageUrl;
-
-    return (
-        <div
-            className={`relative bg-gradient-to-br from-gray-900 via-gray-800 to-gray-900 rounded-2xl shadow-xl border border-gray-700/50 p-6 cursor-pointer transition-all duration-300 hover:scale-[1.02] hover:shadow-2xl hover:border-gray-600/50 ${isHovered ? 'ring-2 ring-purple-500/20' : ''}`}
-            onClick={() => onClick(item)}
-            onMouseEnter={() => setIsHovered(true)}
-            onMouseLeave={() => setIsHovered(false)}
-        >
-            {/* Background accent */}
-            <div className="absolute inset-0 bg-gradient-to-br from-purple-500/5 to-blue-500/5 rounded-2xl" />
-            
-            {currentItemImageStatus?.loading ? (
-                <div className="mb-4 overflow-hidden rounded-xl aspect-video bg-gray-700 flex items-center justify-center">
-                    <div className="w-10 h-10 border-4 border-purple-500 border-t-transparent rounded-full animate-spin"></div>
-                    <span className="text-sm text-gray-400 ml-3">Generating image...</span>
-                </div>
-            ) : imageUrlToDisplay ? (
-                <div className="mb-4 overflow-hidden rounded-xl aspect-video">
-                    <img
-                        src={imageUrlToDisplay}
-                        alt={item.itemName}
-                        className="w-full h-full object-cover transition-transform duration-300 group-hover:scale-105"
-                        onError={(e) => { e.target.onerror = null; e.target.src="https://placehold.co/400x250/333/555?text=Image+Error"; }}
-                    />
-                </div>
-            ) : (
-                <div className="mb-4 overflow-hidden rounded-xl aspect-video bg-gray-700 flex items-center justify-center text-gray-400">
-                    <Camera size={48} className="text-gray-500"/>
-                </div>
-            )}
-
-            <div className="relative flex justify-between items-start">
-                <div className="flex-1">
-                    <h3 className="text-xl font-bold text-gray-100 mb-3 tracking-tight">{item.itemName}</h3>
-                    
-                    <div className="flex items-center gap-4 mb-4">
-                        <div className="flex items-center gap-2">
-                            <StarRating rating={item.rating} size="md" />
-                            <span className="text-lg font-bold text-gray-100">
-                                {item.rating ? item.rating.toFixed(1) : 'N/A'}
-                            </span>
-                        </div>
-                        <div className="flex items-center gap-1.5 text-sm text-gray-400">
-                            <MessageCircle size={14} />
-                            <span>{item.reviewCount || 0} reviews</span>
-                        </div>
-                    </div>
-                    
-                    <div className="flex flex-wrap gap-2">
-                        <span className="inline-flex items-center px-3 py-1.5 rounded-full text-xs font-medium bg-gradient-to-r from-purple-600/20 to-blue-600/20 text-purple-300 border border-purple-500/20 backdrop-blur-sm">
-                            {getCategoryLabel(item.mainCategory, item.category, item.subCategory)}
-                        </span>
-                        {item.location && item.location.city && item.location.city !== "Online" && (
-                            <span className="inline-flex items-center gap-1 px-3 py-1.5 rounded-full text-xs font-medium bg-gradient-to-r from-emerald-600/20 to-teal-600/20 text-emerald-300 border border-emerald-500/20 backdrop-blur-sm">
-                                <MapPin size={12} />
-                                {getLocationLabel(item.location)}
-                            </span>
-                        )}
-                        {item.location && item.location.city === "Online" && (
-                            <span className="inline-flex items-center gap-1 px-3 py-1.5 rounded-full text-xs font-medium bg-gradient-to-r from-blue-600/20 to-cyan-600/20 text-blue-300 border border-blue-500/20 backdrop-blur-sm">
-                                <Globe size={12} />
-                                {language === 'en' ? 'Digital' : 'ดิจิทัล'}
-                            </span>
-                        )}
-                    </div>
-                </div>
-                <div className={`transition-all duration-300 ${isHovered ? 'rotate-90' : ''}`}>
-                    <ChevronDown className="text-gray-500" size={20} />
-                </div>
-            </div>
-
-            {item.reviews && item.reviews.length > 0 && (
-                <div className="mt-4 p-4 bg-gray-800/50 backdrop-blur-sm rounded-xl border border-gray-700/50">
-                    <div className="flex items-center gap-3 mb-2">
-                        <div className="w-8 h-8 bg-gradient-to-br from-purple-500 to-blue-500 rounded-full flex items-center justify-center shadow-lg">
-                            <span className="text-xs font-bold text-white">
-                                {item.reviews[0].author.charAt(0).toUpperCase()}
-                            </span>
-                        </div>
-                        <div className="flex-1">
-                            <div className="flex items-center gap-2">
-                                <span className="text-sm font-semibold text-gray-100">{item.reviews[0].author}</span>
-                                {item.reviews[0].verified && (
-                                    <div className="flex items-center gap-1">
-                                        <CheckCircle className="text-blue-400" size={14} />
-                                        <span className="text-xs text-blue-400">Verified</span>
-                                    </div>
-                                )}
-                            </div>
-                            <StarRating rating={item.reviews[0].rating} size="sm" />
-                        </div>
-                    </div>
-                    <p className="text-sm text-gray-300 line-clamp-2 italic">
-                        "{language === 'en' ? item.reviews[0].comment : item.reviews[0].comment_th || item.reviews[0].comment}"
-                    </p>
-                </div>
-            )}
-        </div>
-    );
-};
-
-// New ItemDetailPage component
-const ItemDetailPage = ({ item, onBack, language, categories, citiesData, bangkokStreetsData, generatedImages, setGeneratedImages }) => {
-
- const currentItemImageStatus = generatedImages[item.id];
-
-// kick off image‐generation when this page mounts (or status changes)
-useEffect(() => {
-  // 1) never run until we have an `item`
-  if (!item) return;
-
-  // 2) only kick off generation if we don’t have a URL
-  //    and we haven’t retried 3× yet
-  if (
-    !currentItemImageStatus ||
-    (
-      !currentItemImageStatus.url &&
-      !currentItemImageStatus.loading &&
-      (
-        currentItemImageStatus.error ||
-        (currentItemImageStatus.retries ?? 0) < 3
-      )
-    )
-  ) {
-    const retries = currentItemImageStatus?.retries ?? 0;
-
-    // mark “loading” and bump retry count
-    setGeneratedImages(prev => ({
-      ...prev,
-      [item.id]: {
-        loading: true,
-        url:     null,
-        error:   false,
-        retries: retries + 1,
-      },
-    }));
-
-    // build your prompt…
-    const prompt = `A detailed photo of the shop front of "${item.itemName}" which is a ${
-      item.subCategory || item.category
-    } in ${
-      item.location.city !== 'Online'
-        ? item.location.city + ', ' + item.location.district
-        : 'the online world'
-    }. Ensure any visible text is in English. Focus on the exterior.`;
-
-    // fire off the API
-    generateImage(prompt)
-      .then(url => {
-        setGeneratedImages(prev => ({
-          ...prev,
-          [item.id]: {
-            ...prev[item.id],
-            loading: false,
-            url,
-            error:   false,
-          }
-        }));
-      })
-      .catch(() => {
-        setGeneratedImages(prev => ({
-          ...prev,
-          [item.id]: {
-            ...prev[item.id],
-            loading: false,
-            url:     null,
-            error:   true,
-          }
-        }));
-      });
-  }
-}, [
-  // deps must include `item` so we bail early when !item
-  item,
-  // track the status fields
-  currentItemImageStatus?.url,
-  currentItemImageStatus?.loading,
-  currentItemImageStatus?.error,
-  currentItemImageStatus?.retries,
-  // and of course the setter
-  setGeneratedImages,
-]);
-
-        const getCategoryLabel = (mainCat, cat, subCat) => {
-        const mainLabel = categories.find(c => c.value === mainCat)?.[`label_${language}`] || mainCat;
-        const platformOrCat = mainCat === 'Online' 
-            ? categories.find(c => c.value === mainCat)?.platforms.find(p => p.value === cat)?.[`label_${language}`] || cat
-            : categories.find(c => c.value === mainCat)?.subcategories.find(s => s.value === cat)?.[`label_${language}`] || cat;
-        const subCatLabel = (mainCat === 'Online' ? (categories.find(c => c.value === mainCat)?.platforms.find(p => p.value === cat)?.subcategories.find(sc => sc.value === subCat)?.[`label_${language}`] || subCat) : (categories.find(c => c.value === mainCat)?.subcategories.find(s => s.value === cat)?.subcategories.find(sc => sc.value === subCat)?.[`label_${language}`] || subCat));
-        
-        let parts = [mainLabel];
-        if (platformOrCat && platformOrCat !== cat) parts.push(platformOrCat);
-        if (subCatLabel && subCatLabel !== subCat) parts.push(subCatLabel);
-
-        if (parts.length === 1 && cat && cat !== 'all') { 
-            const currentCategoryData = (mainCat === 'Online' ? categories.find(c => c.value === mainCat)?.platforms : categories.find(c => c.value === mainCat)?.subcategories)?.find(c => c.value === cat);
-            if (currentCategoryData) parts.push(currentCategoryData[`label_${language}`] || cat);
-        }
-        if (parts.length === 2 && subCat && subCat !== 'all') { 
-            let currentSubCategoryData;
-            if (mainCat === 'Online') {
-                const platform = categories.find(c => c.value === mainCat)?.platforms.find(p => p.value === cat);
-                currentSubCategoryData = platform?.subcategories.find(s => s.value === subCat);
-            } else {
-                const category = categories.find(c => c.value === mainCat)?.subcategories.find(s => s.value === cat);
-                currentSubCategoryData = category?.subcategories.find(sc => sc.value === subCat);
-            }
-            if (currentSubCategoryData) parts.push(currentSubCategoryData[`label_${language}`] || subCat);
-        }
-        return parts.filter(Boolean).join(' / ');
-    };
-
-    const getLocationLabel = (loc) => {
-        if (!loc || !loc.city) return '';
-        const cityLabel = citiesData.find(c => c.value === loc.city)?.[`label_${language}`] || loc.city;
-        const districtLabel = loc.district ? (citiesData.find(c => c.value === loc.city)?.districts.find(d => d.value === loc.district)?.[`label_${language}`] || loc.district) : '';
-        
-        let parts = [cityLabel];
-        if (districtLabel) parts.push(districtLabel);
-
-        return parts.filter(Boolean).join(', ');
-    };
-
-    return (
-        <div className="min-h-screen bg-gradient-to-br from-gray-950 via-gray-900 to-gray-950">
-            <header className="bg-gray-900/90 backdrop-blur-xl shadow-2xl border-b border-gray-800/50 p-4 sticky top-0 z-10">
-                <div className="max-w-4xl mx-auto flex items-center justify-between">
-                    <button onClick={onBack} className="flex items-center gap-2 text-gray-300 hover:text-white transition-colors">
-                        <ArrowLeft size={20} />
-                        <span className="font-semibold">{language === 'en' ? 'Back to List' : 'กลับไปที่รายการ'}</span>
-                    </button>
-                    <h2 className="text-xl font-bold text-gray-100 truncate">{item.itemName}</h2>
-                </div>
-            </header>
-            <main className="max-w-4xl mx-auto p-4 md:p-8">
-                 <div className="bg-gradient-to-br from-gray-900 via-gray-800 to-gray-900 rounded-2xl shadow-xl border border-gray-700/50 p-6 md:p-8 relative overflow-hidden">
-                    {/* Background decoration */}
-                    <div className="absolute top-0 right-0 w-64 h-64 bg-gradient-to-br from-purple-500/10 to-blue-500/10 rounded-full blur-3xl" />
-                    
-                    <div className="relative">
-                        {currentItemImageStatus?.loading ? (
-                            <div className="mb-6 overflow-hidden rounded-xl aspect-video shadow-lg bg-gray-700 flex items-center justify-center">
-                                <div className="w-16 h-16 border-4 border-purple-500 border-t-transparent rounded-full animate-spin"></div>
-                                <span className="text-lg text-gray-400 ml-4">Generating image...</span>
-                            </div>
-                        ) : currentItemImageStatus?.url ? (
-                            <div className="mb-6 overflow-hidden rounded-xl aspect-video shadow-lg">
-                                <img
-                                    src={currentItemImageStatus.url}
-                                    alt={item.itemName}
-                                    className="w-full h-full object-cover"
-                                    onError={(e) => { e.target.onerror = null; e.target.src="https://placehold.co/400x250/333/555?text=Image+Error"; }}
-                                />
-                            </div>
-                        ) : (
-                            <div className="mb-6 overflow-hidden rounded-xl aspect-video shadow-lg bg-gray-700 flex items-center justify-center text-gray-400">
-                                <Camera size={64} className="text-gray-500"/>
-                            </div>
-                        )}
-                        <h2 className="text-3xl font-bold text-gray-100 mb-6 tracking-tight">{item.itemName}</h2>
-                        
-                        <div className="mb-6 flex flex-wrap gap-2">
-                            <span className="inline-flex items-center px-4 py-2 rounded-full text-sm font-medium bg-gradient-to-r from-purple-600/20 to-blue-600/20 text-purple-300 border border-purple-500/20 backdrop-blur-sm">
-                                {getCategoryLabel(item.mainCategory, item.category, item.subCategory)}
-                            </span>
-                            {item.location && item.location.city && item.location.city !== "Online" &&(
-                                <span className="inline-flex items-center gap-1.5 px-4 py-2 rounded-full text-sm font-medium bg-gradient-to-r from-emerald-600/20 to-teal-600/20 text-emerald-300 border border-emerald-500/20 backdrop-blur-sm">
-                                    <MapPin size={14} />
-                                    {getLocationLabel(item.location)}
-                                </span>
-                            )}
-                        </div>
-
-                        <div className="mb-8 p-6 bg-gray-800/50 backdrop-blur-sm rounded-xl border border-gray-700/50">
-                            <div className="flex items-center gap-6">
-                                <div className="text-center">
-                                    <span className="text-5xl font-bold bg-gradient-to-r from-amber-400 to-orange-400 bg-clip-text text-transparent">
-                                        {item.rating ? item.rating.toFixed(1) : 'N/A'}
-                                    </span>
-                                    <p className="text-sm text-gray-400 mt-1">
-                                        {language === 'en' ? 'out of 5' : 'จาก 5'}
-                                    </p>
-                                </div>
-                                <div className="flex-1">
-                                    <StarRating rating={item.rating || 0} size="lg" />
-                                    <p className="text-sm text-gray-400 mt-2">
-                                        {language === 'en' ? `Based on ${item.reviewCount || 0} reviews` : `อิงจาก ${item.reviewCount || 0} รีวิว`}
-                                    </p>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div className="grid grid-cols-2 gap-4 mb-8">
-                            <div className="bg-gradient-to-br from-gray-800/50 to-gray-700/30 backdrop-blur-sm rounded-xl p-4 border border-gray-700/50">
-                                <div className="flex items-center gap-2 text-purple-400 mb-2">
-                                    <TrendingUp size={18} />
-                                    <span className="text-sm font-semibold">{language === 'en' ? "Activity" : "กิจกรรม"}</span>
-                                </div>
-                                <p className="text-lg font-bold text-gray-100">{language === 'en' ? "High" : "สูง"}</p>
-                            </div>
-                            <div className="bg-gradient-to-br from-gray-800/50 to-gray-700/30 backdrop-blur-sm rounded-xl p-4 border border-gray-700/50">
-                                <div className="flex items-center gap-2 text-blue-400 mb-2">
-                                    <Clock size={18} />
-                                    <span className="text-sm font-semibold">{language === 'en' ? "Last Review" : "รีวิวล่าสุด"}</span>
-                                </div>
-                                <p className="text-lg font-bold text-gray-100">{language === 'en' ? "Recent" : "เร็วๆ นี้"}</p>
-                            </div>
-                        </div>
-
-                        <div className="space-y-4">
-                            <h3 className="text-xl font-bold text-gray-100 mb-4">{language === 'en' ? "All Reviews" : "รีวิวทั้งหมด"}</h3>
-                            {item.reviews && item.reviews.length > 0 ? (
-                                item.reviews.sort((a, b) => new Date(b.date) - new Date(a.date)).map((review) => (
-                                    <div key={review.id} className="bg-gray-800/30 backdrop-blur-sm rounded-xl p-5 border border-gray-700/50 hover:border-gray-600/50 transition-all duration-300">
-                                        <div className="flex items-start gap-3 mb-3">
-                                            <div className="w-10 h-10 bg-gradient-to-br from-purple-500 to-blue-500 rounded-full flex items-center justify-center shadow-lg flex-shrink-0">
-                                                <span className="text-sm font-bold text-white">
-                                                    {review.author.charAt(0).toUpperCase()}
-                                                </span>
-                                            </div>
-                                            <div className="flex-1">
-                                                <div className="flex items-center gap-3 mb-1">
-                                                    <span className="font-semibold text-gray-100">{review.author}</span>
-                                                    {review.verified && (
-                                                        <div className="flex items-center gap-1">
-                                                            <CheckCircle className="text-blue-400" size={12} />
-                                                            <span className="text-xs text-blue-400">Verified</span>
-                                                        </div>
-                                                    )}
-                                                    <span className="text-xs text-gray-500">{review.date}</span>
-                                                </div>
-                                                <StarRating rating={review.rating} size="sm" />
-                                            </div>
-                                        </div>
-                                        <p className="text-gray-300 leading-relaxed pl-12">
-                                            {language === 'en' ? review.comment : review.comment_th || review.comment}
-                                        </p>
-                                        <div className="flex items-center gap-4 mt-4 pl-12">
-                                            <button className="flex items-center gap-1.5 text-xs text-gray-400 hover:text-purple-400 transition-colors">
-                                                <Heart size={14} />
-                                                <span>{language === 'en' ? "Helpful" : "มีประโยชน์"} ({review.helpful})</span>
-                                            </button>
-                                        </div>
-                                    </div>
-                                ))
-                            ) : (
-                                <div className="text-center py-8 text-gray-500">
-                                    <MessageCircle className="w-12 h-12 mx-auto mb-3 opacity-50" />
-                                    <p>{language === 'en' ? "No reviews yet. Be the first!" : "ยังไม่มีรีวิว เป็นคนแรก!"}</p>
-                                </div>
-                            )}
-                        </div>
-                    </div>
-                </div>
-            </main>
-        </div>
-    );
-};
 
 
 // Enhanced ReviewFormModal Component
@@ -1388,109 +883,33 @@ const [showSearchBar, setShowSearchBar] = useState(true); // State for search ba
     return (
         <div className="min-h-screen bg-gradient-to-br from-gray-950 via-gray-900 to-gray-950 text-gray-100 font-sans antialiased pt-12">
             {/* Enhanced Header */}
-            <header
-              className={`bg-gray-900/90 backdrop-blur-xl border-b border-gray-800/50 fixed top-0 w-full z-20 transition-transform duration-300 ${showHeader ? "translate-y-0" : "-translate-y-full"}`}
-              style={{ minHeight: "40px", padding: "0.25rem 0" }}
-            >
-              <div className="max-w-4xl mx-auto flex items-center justify-between px-3 py-1 gap-2">
-                <div className="flex items-center gap-2">
-                  <div className="w-7 h-7 bg-gradient-to-br from-purple-600 to-blue-600 rounded-lg flex items-center justify-center">
-                    <Award className="text-white" size={18} />
-                  </div>
-                  <h1 className="text-lg font-black text-transparent bg-clip-text bg-gradient-to-r from-purple-400 to-blue-400 tracking-tight">
-                    ReviewAnything
-                  </h1>
-                </div>
-                <div className="flex items-center gap-1">
-                  <button
-                    onClick={() => setShowSearchBar((prev) => !prev)}
-                    className="p-1 text-gray-400 hover:text-gray-100 hover:bg-gray-800/50 rounded-lg"
-                    title="Search"
-                  >
-                    <Search size={15} />
-                  </button>
-                  <button
-                    onClick={() => setShowFilters((prev) => !prev)}
-                    className="p-1 text-gray-400 hover:text-gray-100 hover:bg-gray-800/50 rounded-lg"
-                    title="Filters"
-                  >
-                    <Filter size={15} />
-                  </button>
-                  <div className="w-px h-5 bg-gray-700 mx-1"></div>
-                  {user ? (
-                    <>
-                      <img src={user.photoURL} alt={user.displayName} className="w-6 h-6 rounded-full" />
-                      <button
-                        onClick={handleLogout}
-                        className="p-1 text-gray-400 hover:text-gray-100 hover:bg-gray-800/50 rounded-lg"
-                        title="Logout"
-                      >
-                        <LogOut size={16} />
-                      </button>
-                    </>
-                  ) : (
-                    <button
-                      onClick={handleLogin}
-                      className="px-2 py-1 bg-white text-gray-800 rounded-lg text-xs shadow hover:shadow-md flex items-center gap-1"
-                    >
-                      <GoogleIcon />
-                      Sign in
-                    </button>
-                  )}
-                  <button
-                    onClick={() => setShowReviewForm(true)}
-                    disabled={!user}
-                    className={`px-2 py-1 bg-gradient-to-r from-purple-600 to-blue-600 text-white rounded-lg text-xs ml-1 ${
-                      !user ? "opacity-50 cursor-not-allowed" : ""
-                    }`}
-                  >
-                    Write
-                  </button>
-                  <button
-                    onClick={() => setLanguage(language === "en" ? "th" : "en")}
-                    className="px-2 py-1 rounded-lg bg-gray-800 text-gray-100 text-xs ml-1"
-                  >
-                    {language === "en" ? "🇹🇭" : "🇬🇧"}
-                  </button>
-                </div>
-              </div>
-            </header>
+            <Header
+                showHeader={showHeader}
+                toggleSearchBar={() => setShowSearchBar(prev => !prev)}
+                toggleFilters={() => setShowFilters(prev => !prev)}
+                user={user}
+                handleLogin={handleLogin}
+                handleLogout={handleLogout}
+                setShowReviewForm={setShowReviewForm}
+                language={language}
+                setLanguage={setLanguage}
+                GoogleIcon={GoogleIcon}
+            />
 
             {/* Main Content */}
-            <main className="max-w-7xl mx-auto px-4 py-8">
-                <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                    {!isAuthReady ? (
-                        <div className="text-center py-20 col-span-full">
-                            <div className="inline-flex items-center justify-center w-16 h-16 bg-gradient-to-br from-purple-600/20 to-blue-600/20 rounded-full mb-4">
-                                <div className="w-10 h-10 border-4 border-purple-500 border-t-transparent rounded-full animate-spin" />
-                            </div>
-                            <p className="text-gray-400">{language === 'en' ? 'Loading Reviews...' : 'กำลังโหลดรีวิว...'}</p>
-                        </div>
-                    ) : filteredReviews.length > 0 ? (
-                        filteredReviews.map((item) => (
-                            <ReviewItem 
-                                key={item.id} 
-                                item={item} 
-                                onClick={handleItemClick} 
-                                language={language} 
-                                categories={categories} 
-                                citiesData={citiesData} 
-                                bangkokStreetsData={bangkokStreetsData} 
-                                generatedImages={generatedImages}
-                                setGeneratedImages={setGeneratedImages}
-                            />
-                        ))
-                    ) : (
-                        <div className="text-center py-20 bg-gradient-to-br from-gray-900 via-gray-800 to-gray-900 rounded-2xl border border-gray-700/50 col-span-full">
-                            <Sparkles className="w-16 h-16 mx-auto mb-4 text-gray-600" />
-                            <p className="text-gray-400 text-lg">
-                                {isLoadingNoMatchingReviewsMessage ? (language === 'en' ? 'Searching...' : 'กำลังค้นหา...') : noMatchingReviewsMessage}
-                            </p>
-                        </div>
-                    )}
-                </div>
-            </main>
-</div>
+            <ItemList
+                filteredReviews={filteredReviews}
+                isAuthReady={isAuthReady}
+                handleItemClick={handleItemClick}
+                language={language}
+                categories={categories}
+                citiesData={citiesData}
+                bangkokStreetsData={bangkokStreetsData}
+                generatedImages={generatedImages}
+                setGeneratedImages={setGeneratedImages}
+                isLoadingNoMatchingReviewsMessage={isLoadingNoMatchingReviewsMessage}
+                noMatchingReviewsMessage={noMatchingReviewsMessage}
+                generateImage={generateImage}
 
 
             <ReviewFormModal
