+++ src/data/sampleReviews.js
@@ -1,0 +1,36 @@
+export const initialSampleReviews = [
+    // Original 9 Items (imageUrl removed, will be generated, fallback added)
+    { id: "sample-1", itemName: "Local Car Repair", mainCategory: "Real World", category: "Local Services", subCategory: "Mechanics", location: { city: "Bangkok", district: "Lat Phrao" }, rating: 4.5, reviewCount: 23, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Car+Repair", reviews: [{ id: 1, author: "Sarah K.", rating: 5, date: "2024-03-15", comment: "Quick and honest car repair.", verified: true, helpful: 15 }] },
+    { id: "sample-2", itemName: "@bangkok_fashion_daily", mainCategory: "Online", category: "Instagram", subCategory: "Influencer", location: { city: "Online" }, rating: 4.2, reviewCount: 60, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Fashion+Blog", reviews: [{ id: 3, author: "Style Seeker", rating: 5, date: "2024-03-20", comment: "Great styling tips!", verified: true, helpful: 32 }] },
+    { id: "sample-3", itemName: "Pro Moving Services", mainCategory: "Real World", category: "Local Services", subCategory: "Movers", location: { city: "Bangkok", district: "Yan Nawa" }, rating: 4.9, reviewCount: 15, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=Moving+Services", reviews: [{ id: 5, author: "Customer C", rating: 5, date: "2024-06-10", comment: "Quick, efficient, and careful.", verified: true, helpful: 7 }] },
+    { id: "sample-4", itemName: "Authentic Isaan Restaurant", mainCategory: "Real World", category: "Restaurants & Cafes", subCategory: "Thai Food", location: { city: "Bangkok", district: "Bang Rak" }, rating: 4.7, reviewCount: 50, fallbackImageUrl: "https://placehold.co/400x250/f39c12/ffffff?text=Isaan+Restaurant", reviews: [{ id: 7, author: "FoodieBKK", rating: 5, date: "2024-05-10", comment: "Amazing Som Tum and Larb Gai!", verified: true, helpful: 25 }] },
+    { id: "sample-5", itemName: "Math Tutor Expert", mainCategory: "Real World", category: "Education Centers", subCategory: "Tutoring Center", location: { city: "Bangkok", district: "Pathum Wan" }, rating: 4.5, reviewCount: 10, fallbackImageUrl: "https://placehold.co/400x250/9b59b6/ffffff?text=Math+Tutor", reviews: [{ id: 8, author: "Student Parent", rating: 5, date: "2024-05-20", comment: "Our son's grades improved significantly.", verified: true, helpful: 5 }] },
+    { id: "sample-6", itemName: "Vintage Camera Shop", mainCategory: "Real World", category: "Goods & Products", subCategory: "Collectibles", location: { city: "Bangkok", district: "Bang Rak" }, rating: 4.0, reviewCount: 12, fallbackImageUrl: "https://placehold.co/400x250/1abc9c/ffffff?text=Camera+Shop", reviews: [{ id: 9, author: "CollectorGuy", rating: 4, date: "2024-06-01", comment: "Found a rare vintage camera.", verified: false, helpful: 3 }] },
+    { id: "sample-7", itemName: "Online Marketing Agency", mainCategory: "Online", category: "Website", subCategory: "Service Provider", location: { city: "Online" }, rating: 4.7, reviewCount: 8, fallbackImageUrl: "https://placehold.co/400x250/34495e/ffffff?text=Marketing+Agency", reviews: [{ id: 10, author: "Client M", rating: 5, date: "2024-06-15", comment: "Amazing post-processing work.", verified: true, helpful: 2 }] },
+    { id: "sample-8", itemName: "Hip Cafe", mainCategory: "Real World", category: "Restaurants & Cafes", subCategory: "Cafe", location: { city: "Bangkok", district: "Watthana" }, rating: 4.3, reviewCount: 30, fallbackImageUrl: "https://placehold.co/400x250/95a5a6/ffffff?text=Hip+Cafe", reviews: [{ id: 11, author: "Coffee Lover", rating: 5, date: "2024-06-18", comment: "Best pour-over coffee in Thong Lo.", verified: true, helpful: 10 }] },
+    { id: "sample-9", itemName: "Co-working Space", mainCategory: "Real World", category: "Local Services", subCategory: "Other", location: { city: "Bangkok", district: "Watthana" }, rating: 4.6, reviewCount: 25, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Co-working", reviews: [{ id: 12, author: "Digital Nomad", rating: 5, date: "2024-06-20", comment: "Fast internet, comfy chairs.", verified: true, helpful: 8 }] },
+
+    // --- NEWLY ADDED ITEMS FOR FULL SUB-CATEGORY COVERAGE ---
+    { id: "new-fb-shop", itemName: "BKK Gadgets (FB)", mainCategory: "Online", category: "Facebook", subCategory: "Shop", location: { city: "Online" }, rating: 4.8, reviewCount: 150, fallbackImageUrl: "https://placehold.co/400x250/2980b9/ffffff?text=Gadgets+Shop", reviews: [{ id: 13, author: "TechieTom", rating: 5, date: "2024-06-12", comment: "Fast shipping on Facebook.", verified: true, helpful: 45 }] },
+    { id: "new-fb-page", itemName: "Thai Food Lovers Page", mainCategory: "Online", category: "Facebook", subCategory: "Page", location: { city: "Online" }, rating: 4.9, reviewCount: 1200, fallbackImageUrl: "https://placehold.co/400x250/8e44ad/ffffff?text=Food+Page", reviews: [{ id: 14, author: "Somchai C.", rating: 5, date: "2024-05-30", comment: "Best recipes.", verified: true, helpful: 250 }] },
+    { id: "new-fb-group", itemName: "Bangkok Expat Buy/Sell", mainCategory: "Online", category: "Facebook", subCategory: "Group", location: { city: "Online" }, rating: 4.5, reviewCount: 45, fallbackImageUrl: "https://placehold.co/400x250/2c3e50/ffffff?text=Buy/Sell+Group", reviews: [{ id: 15, author: "NewbieNancy", rating: 4, date: "2024-06-10", comment: "Found great deals.", verified: false, helpful: 12 }] },
+    { id: "new-shopee-shop", itemName: "SuperDeals Shopee", mainCategory: "Online", category: "Shopee", subCategory: "Shop", location: { city: "Online" }, rating: 4.7, reviewCount: 5000, fallbackImageUrl: "https://placehold.co/400x250/f1c40f/ffffff?text=Shopee+Shop", reviews: [{ id: 16, author: "ShopperSue", rating: 5, date: "2024-06-14", comment: "My go-to shop on Shopee.", verified: true, helpful: 123 }] },
+    { id: "new-shopee-seller", itemName: "Seller: JaneD_bkk", mainCategory: "Online", category: "Shopee", subCategory: "Seller", location: { city: "Online" }, rating: 4.9, reviewCount: 210, fallbackImageUrl: "https://placehold.co/400x250/e67e22/ffffff?text=Shopee+Seller", reviews: [{ id: 17, author: "HappyBuyer", rating: 5, date: "2024-06-11", comment: "Seller was very responsive.", verified: true, helpful: 30 }] },
+    { id: "new-ig-shop", itemName: "@ChicBoutique.BKK", mainCategory: "Online", category: "Instagram", subCategory: "Shop", location: { city: "Online" }, rating: 4.6, reviewCount: 88, fallbackImageUrl: "https://placehold.co/400x250/16a085/ffffff?text=Chic+Boutique", reviews: [{ id: 18, author: "FashionistaFae", rating: 5, date: "2024-06-05", comment: "The dress is gorgeous!", verified: true, helpful: 22 }] },
+    { id: "new-tiktok-influencer", itemName: "DanceMachine_TikTok", mainCategory: "Online", category: "TikTok", subCategory: "Influencer", location: { city: "Online" }, rating: 4.9, reviewCount: 1200, fallbackImageUrl: "https://placehold.co/400x250/27ae60/ffffff?text=TikTok+Influencer", reviews: [{ id: 19, author: "Fan1", rating: 5, date: "2025-06-17", comment: "Best dance challenges!", verified: false, helpful: 300 }] },
+    { id: "new-tiktok-shop", itemName: "TikTok Finds Store", mainCategory: "Online", category: "TikTok", subCategory: "Shop", location: { city: "Online" }, rating: 4.5, reviewCount: 120, fallbackImageUrl: "https://placehold.co/400x250/d35400/ffffff?text=TikTok+Shop", reviews: [{ id: 20, author: "ViralBuyer", rating: 4, date: "2024-06-30", comment: "Found the viral gadget here.", verified: true, helpful: 50 }] },
+    { id: "new-yt-channel", itemName: "Tech Tomorrow", mainCategory: "Online", category: "YouTube", subCategory: "Channel", location: { city: "Online" }, rating: 4.8, reviewCount: 450, fallbackImageUrl: "https://placehold.co/400x250/c0392b/ffffff?text=YouTube+Channel", reviews: [{ id: 21, author: "GadgetGeek", rating: 5, date: "2024-06-28", comment: "Unbiased and detailed reviews.", verified: true, helpful: 150 }] },
+    { id: "new-yt-creator", itemName: "Creator: Chef Praiya", mainCategory: "Online", category: "YouTube", subCategory: "Content Creator", location: { city: "Online" }, rating: 4.9, reviewCount: 800, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Chef+Creator", reviews: [{ id: 22, author: "HomeCook", rating: 5, date: "2024-06-25", comment: "Her cooking is so inspiring.", verified: true, helpful: 200 }] },
+    { id: "new-ls-tutor", itemName: "English Express Tutors", mainCategory: "Real World", category: "Local Services", subCategory: "Tutors", location: { city: "Bangkok", district: "Pathum Wan" }, rating: 4.7, reviewCount: 35, fallbackImageUrl: "https://placehold.co/400x250/1abc9c/ffffff?text=English+Tutor", reviews: [{ id: 23, author: "Parent_P", rating: 5, date: "2024-06-20", comment: "My daughter's confidence in English has soared.", verified: true, helpful: 10 }] },
+    { id: "new-ls-handyman", itemName: "Mr. Fix-It", mainCategory: "Real World", category: "Local Services", subCategory: "Handymen", location: { city: "Bangkok", district: "Watthana" }, rating: 4.8, reviewCount: 42, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=Handyman", reviews: [{ id: 24, author: "HarryH", rating: 5, date: "2024-06-01", comment: "Fixed my leaky faucet in no time.", verified: true, helpful: 18 }] },
+    { id: "new-ls-cleaner", itemName: "SparkleClean BKK", mainCategory: "Real World", category: "Local Services", subCategory: "Cleaners", location: { city: "Bangkok", district: "Sathorn" }, rating: 4.9, reviewCount: 75, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Cleaner", reviews: [{ id: 25, author: "BusyBob", rating: 5, date: "2024-06-15", comment: "My condo has never been this clean.", verified: true, helpful: 25 }] },
+    { id: "new-ls-plumber", itemName: "Reliable Plumbers", mainCategory: "Real World", category: "Local Services", subCategory: "Plumbers", location: { city: "Bangkok", district: "Chatuchak" }, rating: 4.7, reviewCount: 33, fallbackImageUrl: "https://placehold.co/400x250/9b59b6/ffffff?text=Plumber", reviews: [{ id: 26, author: "StressedSteve", rating: 5, date: "2024-05-25", comment: "Came at 2 AM for a burst pipe. Lifesavers!", verified: true, helpful: 11 }] },
+    { id: "new-ls-electrician", itemName: "Watt's Up Electricians", mainCategory: "Real World", category: "Local Services", subCategory: "Electricians", location: { city: "Bangkok", district: "Bang Kapi" }, rating: 4.8, reviewCount: 29, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Electrician", reviews: [{ id: 27, author: "OfficeOlly", rating: 5, date: "2024-06-02", comment: "Rewired our office over the weekend.", verified: true, helpful: 9 }] },
+    { id: "new-fb-car", itemName: "John's Used Cars (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Cars", location: { city: "Bangkok", district: "Bang Khen" }, rating: 4.4, reviewCount: 19, fallbackImageUrl: "https://placehold.co/400x250/f39c12/ffffff?text=Used+Car", reviews: [{ id: 28, author: "FrankF", rating: 4, date: "2024-06-08", comment: "Car was as described. Fair deal.", verified: false, helpful: 5 }] },
+    { id: "new-fb-furniture", itemName: "Secondhand Sofa (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Furniture", location: { city: "Bangkok", district: "Prawet" }, rating: 5.0, reviewCount: 1, fallbackImageUrl: "https://placehold.co/400x250/7f8c8d/ffffff?text=Used+Sofa", reviews: [{ id: 29, author: "HappyHome", rating: 5, date: "2025-06-17", comment: "Great condition and a great price!", verified: false, helpful: 1 }] },
+    { id: "new-fb-electronics", itemName: "Used iPhone (FB)", mainCategory: "Real World", category: "Facebook Marketplace", subCategory: "Used Electronics", location: { city: "Bangkok", district: "Don Mueang" }, rating: 4.0, reviewCount: 5, fallbackImageUrl: "https://placehold.co/400x250/2c3e50/ffffff?text=Used+iPhone", reviews: [{ id: 30, author: "BuyerBen", rating: 4, date: "2024-06-29", comment: "Battery health was lower than stated, but works.", verified: false, helpful: 2 }] },
+    { id: "new-ll-condo", itemName: "Landlord: Mrs. Somchai", mainCategory: "Real World", category: "Landlords", subCategory: "Condo", location: { city: "Bangkok", district: "Huai Khwang" }, rating: 4.9, reviewCount: 8, fallbackImageUrl: "https://placehold.co/400x250/3498db/ffffff?text=Condo+Building", reviews: [{ id: 31, author: "TenantTom", rating: 5, date: "2024-06-01", comment: "Best landlord ever.", verified: true, helpful: 4 }] },
+    { id: "new-ll-apt", itemName: "Sukhumvit Suites", mainCategory: "Real World", category: "Landlords", subCategory: "Apartment", location: { city: "Bangkok", district: "Khlong Toei" }, rating: 4.1, reviewCount: 22, fallbackImageUrl: "https://placehold.co/400x250/e74c3c/ffffff?text=Apartment+Building", reviews: [{ id: 32, author: "ExpatEve", rating: 4, date: "2024-05-15", comment: "Good location, but the walls are a bit thin.", verified: true, helpful: 6 }] },
+    { id: "new-ll-house", itemName: "Rental House Ladprao", mainCategory: "Real World", category: "Landlords", subCategory: "House", location: { city: "Bangkok", district: "Lat Phrao" }, rating: 4.6, reviewCount: 3, fallbackImageUrl: "https://placehold.co/400x250/2ecc71/ffffff?text=House+Rental", reviews: [{ id: 33, author: "FamilyMan", rating: 5, date: "2024-04-10", comment: "Perfect for our family, great garden.", verified: true, helpful: 1 }] },
+];
+
